Hopefully prevent buffer overflow of error message buffer.
The way error messages are handled in this tool is fairly sketchy and
overall pretty cumbersome. This change increases the size of emsg[]
to account for some overlooked cases where the entirety of tmp[] could
be substituted into emsg[].

I don't much like this 'fix'; redoing the error message handling would
be a much better idea as sane implementations would eliminate the
emsg[] buffer concept entirely. However, until that were to occur, this
change is certainly better than nothing.